Move motionManager to MotionDataStore
